move two queries from CourseList.js to CourseSearch.js and refine CourseList props
